+++ src/borneo/driver.py
@@ -420,7 +420,7 @@ class NoSQLHandle(object):
             an instance of :py:class:`QueryRequest`.
         :raises NoSQLException: raises the exception if the operation cannot be
             performed for any other reason.
-        :versionadded: 5.3.5
+        :versionadded: 5.3.6
         """
         if not isinstance(request, QueryRequest):
             raise IllegalArgumentException(