+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Added
 _____
 
 * Cloud only: New regions: us-tacoma-1, us-chicago-1, eu-dcc-madrid-1.
+* Added new method borneo.NoSQLHandle.query_iterable(query_request) to get an iterable
+that contains all the results of a query.
 
 ====================
  5.3.5 - 2022-08-09